scrapers/requirements.txt
osmnx>=1.7.0
supabase>=2.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
populartimes>=0.0.0
